requirements_eval.txt
absl-py==1.2.0
addict==2.4.0
ai2thor==2.1.0
antlr4-python3-runtime==4.9.3
anyio==3.6.1
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
async-generator==1.10
attrs==22.1.0
aws-requests-auth==0.4.3
Babel==2.10.3
backcall==0.2.0
beautifulsoup4==4.11.1
black==21.4b2
bleach==4.1.0
boto3==1.23.10
botocore==1.26.10
cached-property==1.5.2
cachetools==4.2.4
certifi==2021.5.30
cffi==1.15.1
charset-normalizer==2.0.12
click==8.0.4
clip @ git+https://github.com/openai/CLIP.git@d50d76daa670286dd6cacf3bcd80b5e4823fc8e1
cloudpickle==2.1.0
colorama==0.4.5
contextvars==2.4
cycler==0.11.0
dataclasses==0.8
decorator==4.4.2
defusedxml==0.7.1
deprecation==2.1.0
docopt==0.6.2
easydict==1.10
einops==0.4.1
entrypoints==0.4
filelock==3.4.1
Flask==2.0.3
ftfy==6.0.3
future==0.18.2
fvcore==0.1.5.post20220512
gdown==4.5.3
gitdb==4.0.9
GitPython==3.1.18
google-auth==2.11.0
google-auth-oauthlib==0.4.6
grpcio==1.47.0
gtimer==1.0.0b5
h5py==3.1.0
huggingface-hub==0.4.0
hydra-core==1.2.0
idna==3.3
imageio==2.15.0
immutables==0.18
importlib-metadata==4.8.3
importlib-resources==5.4.0
iopath==0.1.9
ipykernel==5.5.6
ipython==7.16.3
ipython-genutils==0.2.0
ipywidgets==7.7.2
itsdangerous==2.0.1
jedi==0.17.2
Jinja2==3.0.3
jmespath==0.10.0
joblib==1.1.0
json5==0.9.10
jsonpickle==1.5.2
jsonschema==3.2.0
jupyter-client==7.1.2
jupyter-core==4.9.2
jupyter-packaging==0.10.6
jupyter-server==1.13.1
jupyterlab==3.2.9
jupyterlab-pygments==0.1.2
jupyterlab-server==2.10.3
jupyterlab-widgets==1.1.1
kiwisolver==1.3.1
lmdb==1.3.0
Markdown==3.3.7
MarkupSafe==2.0.1
matplotlib==3.3.4
mistune==0.8.4
msgpack==1.0.4
munch==2.5.0
mypy-extensions==0.4.3
nbclassic==0.3.5
nbclient==0.5.9
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.5
networkx==2.5.1
nltk==3.6.7
notebook==6.4.10
numpy==1.19.5
oauthlib==3.2.0
omegaconf==2.2.3
openai==0.8.0
opencv-python==4.6.0.66
packaging==21.3
pandas==1.1.5
pandocfilters==1.5.0
parso==0.7.1
pathspec==0.9.0
pexpect==4.8.0
pickle5==0.0.12
pickleshare==0.7.5
Pillow==8.4.0
portalocker==2.5.1
progressbar2==3.55.0
prometheus-client==0.14.1
prompt-toolkit==3.0.30
protobuf==3.19.4
ptyprocess==0.7.0
py-cpuinfo==8.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.4
pycparser==2.21
pydot==1.4.2
Pygments==2.13.0
pyparsing==3.0.9
pyquaternion==0.9.9
pyrsistent==0.18.0
PySocks==1.7.1
python-dateutil==2.8.2
python-utils==3.3.3
python-xlib==0.32
pytorch-ranger==0.1.1
pytorch-transformers==1.2.0
pytz==2022.2
PyWavelets==1.1.1
PyYAML==6.0
pyzmq==23.2.1
randaugment==1.0.2
regex==2022.8.17
requests==2.27.1
requests-oauthlib==1.3.1
revtok==0.0.3
rsa==4.9
s3transfer==0.5.2
sacred==0.8.2
sacremoses==0.0.53
scikit-image==0.17.2
scikit-learn==0.24.2
scipy==1.5.4
seaborn==0.11.2
Send2Trash==1.8.0
sentence-transformers==2.2.2
sentencepiece==0.1.97
six==1.16.0
sklearn==0.0
smmap==5.0.0
sniffio==1.2.0
soupsieve==2.3.2.post1
tabulate==0.8.10
tensorboard==2.10.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==1.8
termcolor==1.1.0
terminado==0.12.1
testpath==0.6.0
threadpoolctl==3.1.0
tifffile==2020.9.3
tokenizers==0.12.1
toml==0.10.2
tomlkit==0.11.4
tornado==6.1
tqdm==4.64.0
traitlets==4.3.3
transformers==4.18.0
typed-ast==1.5.4
typing_extensions==4.1.1
urllib3==1.26.11
vit-pytorch==0.38.1
vocab==0.0.5
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.3.1
Werkzeug==2.0.3
widgetsnbextension==3.6.1
wrapt==1.14.1
yacs==0.1.8
zipp==3.6.0
